requirements.txt
# === Core Framework ===
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# === Auth ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# === Database (Neon Postgres) ===
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0

# === Cloudflare R2 (S3-compatible) ===
boto3==1.34.11

# === HTTP Client (Upstash REST + general) ===
httpx==0.25.2

# === Google Cloud ===
google-cloud-secret-manager==2.18.1
google-cloud-documentai==2.24.0
google-cloud-logging==3.8.0

# === Firebase (Server-Side Push Notifications) ===
firebase-admin==6.4.0

# === Integrations ===
# sendgrid removed in favor of brevo (via httpx)
stripe==7.10.0

# === Monitoring ===
structlog==23.2.0

# === Security ===
cryptography==41.0.7

# === Testing ===
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
factory-boy==3.3.1

# === Utilities ===
python-dateutil==2.8.2
orjson==3.9.10
tenacity==8.2.3
